refactor(club): /simplify pass — reuse shared SortableHeader, move tier map to distinguishedTier, dedupe EM_DASH (#1229)
