+++ frontend/src/utils/__tests__/distinguishedTier.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { distinguishedTierName } from '../distinguishedTier'
+import { distinguishedTierName, normalizeTierCode } from '../distinguishedTier'
 
 /* #795 (epic #821 Sprint 3) — domain helper for rendering the "Club
    Distinguished Status" code (`'' | D | S | P | M`) as a display name.
@@ -22,5 +22,34 @@ describe('distinguishedTierName (#795)', () => {
 
   it('falls back to "Distinguished" for an unknown non-empty code', () => {
     expect(distinguishedTierName('Z')).toBe('Distinguished')
+  })
+})
+
+describe('normalizeTierCode (#1229)', () => {
+  it('returns null for absent / empty status (no distinguished status)', () => {
+    expect(normalizeTierCode(undefined)).toBeNull()
+    expect(normalizeTierCode('')).toBeNull()
+  })
+
+  it('passes through canonical letter codes', () => {
+    expect(normalizeTierCode('D')).toBe('D')
+    expect(normalizeTierCode('S')).toBe('S')
+    expect(normalizeTierCode('P')).toBe('P')
+    expect(normalizeTierCode('M')).toBe('M')
+  })
+
+  it('maps historical word forms back to letter codes (incl. Smedley)', () => {
+    expect(normalizeTierCode('Distinguished')).toBe('D')
+    expect(normalizeTierCode('Select Distinguished')).toBe('S')
+    expect(normalizeTierCode("President's Distinguished")).toBe('P')
+    expect(normalizeTierCode('Smedley Distinguished')).toBe('M')
+  })
+
+  it('is case- and whitespace-insensitive for word forms', () => {
+    expect(normalizeTierCode('  smedley distinguished  ')).toBe('M')
+  })
+
+  it('returns null for an unrecognised value rather than guessing', () => {
+    expect(normalizeTierCode('Mystery')).toBeNull()
   })
 })
